backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Redis and Celery
redis==5.0.1
celery==5.3.4

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyjwt==2.8.0

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP client
httpx==0.25.2
requests==2.31.0

# AWS SDK
boto3==1.34.0

# Google Earth Engine
earthengine-api==0.1.384

# Image processing
Pillow==10.1.0

# PDF generation
reportlab==4.0.7

# SMS (Africa's Talking)
africastalking==1.2.6

# Environment variables
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
PyPDF2==3.0.1

# Code quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Monitoring
sentry-sdk[fastapi]==1.38.0
